rename fp mod
