+++ finality-gadget/operator/finalityprovider/app.go
@@ -1,4 +1,4 @@
-package fp
+package finalityprovider
 
 import (
 	"context"
@@ -17,7 +17,7 @@ import (
 
 	"github.com/alt-research/blitz/finality-gadget/client/eotsmanager"
 	"github.com/alt-research/blitz/finality-gadget/operator/configs"
-	"github.com/alt-research/blitz/finality-gadget/operator/fp/controllers"
+	"github.com/alt-research/blitz/finality-gadget/operator/finalityprovider/controllers"
 )
 
 type FinalityProviderApp struct {
